+++ Depense/Model/MonLieu.cs
@@ -13,7 +13,7 @@ namespace Depense.Model
         public string Nom { get; set; }
         public string Adresse { get; set; }
         public string Categorie { get; set; }
-        public Decimal Latitude { get; set; }
-        public Decimal Longitude { get; set; }
+        public Double Latitude { get; set; }
+        public Double Longitude { get; set; }
     }
 }
